+++ projects/16-open-qf/src/site/components/project/info.jsx
