+++ projects/16-open-qf/src/site/pages/project/[id].jsx
@@ -2,9 +2,9 @@ import ListLayout from "@/components/layouts/listLayout";
 import { ROUND_LIST_DATA } from "@/fixtures/roundList";
 import { find } from "lodash-es";
 import { useRouter } from "next/router";
-import ExploreProjectInfo from "@/components/explore/info";
+import ExploreProjectInfo from "@/components/project/info";
 
-export default function ExplorePage() {
+export default function ProjectPage() {
   const router = useRouter();
   const id = Number(router.query.id);
